Refactored to use react-redux's connect() to generate container components.
