+++ .eslintrc
@@ -1,4 +1,7 @@
 {
   "extends": ["standard", "standard-react"],
-  "parser": "babel-eslint"
+  "parser": "babel-eslint",
+  "rules": {
+    "react/prop-types": 0
+  }
 }
