+++ src/components/Footer.js
@@ -1,13 +1,15 @@
 import React from 'react'
-import FilterLink from './FilterLink'
+import Link from './Link'
 
 const Footer = () => {
   return (
     <p>
       Show:
-      <FilterLink filter="SHOW_ALL">All</FilterLink>{' '}
-      <FilterLink filter="SHOW_ACTIVE">Active</FilterLink>{' '}
-      <FilterLink filter="SHOW_COMPLETED">Completed</FilterLink>
+      <Link filter='SHOW_ALL'>All</Link>
+      {','}
+      <Link filter='SHOW_ACTIVE'>Active</Link>
+      {','}
+      <Link filter='SHOW_COMPLETED'>Completed</Link>
     </p>
   )
 }
